Refactor 'nose.importer' to replace 'imp' functions with 'importlib'
This commit refactors the `nose/importer.py` module to replace usage of the `imp`
module with the recommended `importlib`, as the former is has been deprecated since
Python 3.4 and expected to be removed in Python 3.12.

The `find_module` and `load_module` functions have been replaced with local functions
that define similar behavior (though not fully integrated with compiled modules).
However, for plain source files, builtins, and package folders it should work the same.

Some of the code has been re-implemented using the internal `_imp` module, which could
potentially be removed with `imp` (but does not emit an import warning).
This includes the `acquire_lock` and `release_lock` methods originally imported through `imp`.

Bumped package version to '1.3.7.1'

https://github.com/nose-devs/nose/issues/559